Kafka zipkin collector
